[FEATURE] Condition and Condition / Extend ViewHelpers
Here's a fun one for you all:

An extended condition ViewHelper for those hard-to-reach spots. It
supports complex syntax such as:

({foo} == {bar} AND {baz} == 0) OR {foz} < 100)

And the expression is then parsed recursively, analyzing each group of
expressions.

The current version is somewhere around 50 times slower than f:if
because of the need to use a nested ViewHelper to support the
expression. There are optimizations to be achieved still.

Example syntax:

    <v:condition>
        <v:condition.extend>({a} == {b} AND {c} > 0) OR {d} <
100</v:condition.extend>
        This text is ignored; only the f:then / f:else and "then" and
"else" arguments are respected
        <f:then>Success!</f:then>
        <f:else>Tough luck, amigo.</f:else>
    </v:condition>