[FEATURE] Form / Required ViewHelper
Condition ViewHelper to render the THEN child if the $property given is required by a @validate NotEmpty (the particular Validator to trigger a required TRUE can be overridden). Supports dotted path syntax to the property to be checked.

Useful to pre-indicate a field as required without hardcoding a check for field name or similar workarounds.