[FEATURE] Merge support, AbstractAssetViewHelper
Adds merging of Assets of the same type as long as assets are not marked "standalone". If a group of interdependent Assets contains one standalone Asset in the middle of the dependency list, two merged files and one standalone will be generated, i.e.:

1. chunk of merged code up until the standalone Asset
2. a single standalone Asset inclusion
3. the remaining chunk of merged code after the standalone Asset

Any number of standalone Assets may be introduced. Standalone status can be set through TypoScript on groups of Assets or individual Assets. Dependency order will be respected regardless.