[FEATURE] Add "external" attribute on AbstractAssetViewHelper
Setting this boolean flag to TRUE causes the Asset to be considered external, remove.

If an Asset is added by path and the path is remote, but "external" is NOT enabled, the Asset source is downloaded and merged with the other files. Which means that for CDN sources you simply **must** use "external" (also when, for example, overriding paths in TypoScript to use CDN sources).