[BUGFIX] force close tag "script"
Tag "script" must be closed for not messing up the page. Actually, the
self-closing XML syntax <script /> is correct XML, but for it to work
in practice, the web server also needs to send the XML mime-type
like application/xhtml+xml and not as text/html which is not something
we want by default.