[FEATURE] Deferred menu rendering ViewHelper
This adds attribute "deferred" (boolean) to ever menu-type ViewHelper. If set to TRUE, the menu then requires an instance of v:page.menu.deferred to be used inside the actual menu tag. This sub-ViewHelper then outputs the actual menu as rendered by the parent instance (which means you can now both rely on automatic rendering AND perform checks on the menu array which before, you would not be able to do unless performing manual menu rendering).