[TASK] Use valid HTML comment delimiters (finally)
The previous commit on this (846e418) only removed the '!' char in the
comment end delimiter, but did not actually adjust the comment delimiters
to meet the HTML comment syntax as described by the W3C
(http://www.w3.org/TR/html-markup/syntax.html#comments) as stated in the
commit message.

This commit finally does by adjusting the delimiters to the usual syntax
('<!-- ... -->').