[BUGFIX] AbstractContentVH returned content record AND the translation
This change fixes an issue where the method getContentRecords would return
the original and the translated content record if the language of the
original was set to all and a translation for the current language was
provided.

Fixed by selecting only content records with no visible translation for
the current language.