[FEATURE] Chunk ViewHelper
array_chunk() is a really helpful function which functionality should be available as a ViewHelper.

Fx if you have an Array of pages from v:page.menu and want to allocate these links to a fixed number of container elements this would be much code without this ViewHelper.

Fx we have 9 links which shall be placed in 2 containers of same count of items each, if odd the first container takes more all.

Expected html markup:
<div class="container1">
<a>1. Link</a>
<a>2. Link</a>
<a>3. Link</a>
<a>4. Link</a>
<a>5. Link</a>
</div>
<div class="container2">
<a>6. Link</a>
<a>7. Link</a>
<a>8. Link</a>
<a>9. Link</a>
</div>

With this ViewHelper this is an easy task:
{links->v:iterator.chunk(count:2,fixed:1)->v:var.set(name:'chunks')}
<div class="container1">
<f:for each="{chunks.0}" as="item">{item}</f:for>
</div>
<div class="container2">
<f:for each="{chunks.1}" as="item">{item}</f:for>
</div>