[TASK] Removing unused argument "bullet"
The argument bullet designed to insert html before each item, but this is a rare use case especially because there are alternatives fx by css or manual rendering.

Since the functionality for this argument is not yet implemented we will remove the argument.

Who needs something like that should use the manual rendering instead.