[FEATURE] Cookie-less domain prefix setting
This commit adds a `plugin.tx_vhs.settings.prependPath` setting which when used is inserted in front of all (relative) asset paths - including those returned by VHS's image ViewHelper and relative resource path ViewHelpers. In addition, the ViewHelper `v:page.prefixPath` ViewHelper is added which outputs this setting's value so it can be used with other ViewHelpers like `f:uri.image` or `f:uri.resource`.

This setting has precedence over VHS's respect for `config.absRefPrefix` which means it is capable of containing a secondary domain name and protocol, allowing all URIs built for assets to _use a cookie-less domain name for all asset paths_. Doing this is generally considered a good idea since it allows requests to that domain to set and receive no cookies and thus be smaller and faster to process.