[BUGFIX] Complete array returned for fileReference
When we replaced toArray with getProperties to get file reference the properties of the File itself were no longer available because toArray performed a array_merge internally.

For that we created a own method getFileArray but didn't use this in content.resources.fal yet.

This bugfix moves that method to a utility and replaces all usages of that method to use it on the utility.

Fixes: #532