[BUGFIX] Avoid path conflicts on CSS assets
When copying files referenced inside CSS files (the "rewrite" process),
only the filename of the examined CSS file is taken into account for
obtaining a temporary name. That leads to name conflicts if two CSS
files both reference to a – different – file that is reachable via the
same relative path from within each of the two CSS files.