[TASK] Add legacy support for Asset "allowMoveToFooter"
This adds a remapping of allowMoveToFooter->movable if Asset is created from an array, which happens when Asset is defined in TypoScript. The PHP API for Asset has always required the "movable" attribute but documentation for TypoScript settings named the "allowMoveToFooter" attribute instead.

A deprecation log message will be dispatched if the old attribute is encountered, but the remapping means no breakage should occur.

Remapping is to be removed in 2.2 or 3.0, whichever comes first.