[FEATURE] Implement reusable ViewHelperTraits
This change adds:

* BasicViewHelperTrait
* ArrayConsumingViewHelperTrait
* TemplateVariableViewHelperTrait

And removes:

* ViewHelperUtility

Each Trait contains special methods which make it easier to consistently perform repeated actions common to many ViewHelpers - for example: getting an argument either from argument name or tag content, getting an argument and ensuring it gets converted to an array, rendering the tag contents with custom variables that are backed up if they exist, and more.

The Traits are then implemented in our current ViewHelpers to remove a lot of duplicated code. In addition, the former static calls are avoided, making mocking of tests easier.