[FEATURE] Change behaviour of getSlideRecordsFromPage from AbstractContentViewHelper - Add pageSelectService to SlideViewHelperTrait - Use hidePageForLanguageUid from PageSelectService
Why this is done?
If you have a content_fallback defined and want to render a content from a **different** page
than the one displayed with vhs:content.render viewhelper, it won't work in the way expected.
Instead it uses the sys_language_content value of the current page and not the sys_language_content
of the page which the content is on.