[BUGFIX] Fixed deprecation warnings with TSFE->getPageRenderer()
As stated at

https://wiki.typo3.org/TYPO3.CMS/Releases/7.4/Deprecation#Deprecation:_.2368074_-_Deprecate_getPageRenderer.28.29_methods

the getPageRenderer()-function has been marked as deprecated with the release of TYPO3 7.4 and will be removed with the
release of TYPO3 8.0. PageRenderer implements SingletonInterface, which means that getting a own (shared) instance of it will become mandatory.

Therefore, in favor of DRY-principle, a Trait has been introduced providing a instance of PageRenderer through GeneralUtility-instanciation.