[FEATURE] Get all settings from v:variable.extensionConfiguration
By passing an empty path or leaving out the path argument the entire configuration array of the extension is now returned. In addition an internal cache is implemented to avoid unserializing and removing TS dots from arrays when not required - and the ViewHelper class is refactored to use registerArguments and renderStatic for performance.