[FEATURE] Image canvas feature
This change adds a new feature to v:media.image to create a canvas with given dimensions and background color to center an image on. This is useful in cases where images of different sizes and aspect ratios need to be displayed in a clean way without having to crop them.

Example: `<v:media.image src="{image}" alt="" width="600" canvasWidth="800" canvasHeight="500" canvasColor="#cccccc" />`