[FEATURE] Add crop feature to media.image srcset
Division by zero check when calculating ratio
Wrap can be ignored when $treatIdAsReference=TRUE
Height calculation can be ignored when height attribute is missing

When $treatIdAsReference = FALSE the generation of the srcset images does not work. There is a problem when processed images are getting processed again.

If $treatIdAsReference = TRUE the srcset images are getting processed from the original image. Therefore the crop and ratio information is needed.