[TASK] Preparations for LTS-v8 compiling support bridge (#1151)
This patch adds a couple of changes which will be required for an in-progress conversion of every ViewHelper which can support it, to ideal performance when compiled.

A new dependency, namelesscoder/typo3-cms-fluid-gap, provides the necessary API to achieve this. Slight rewriting of the test base ensures all ViewHelpers can be tested with
proxying from render() to renderStatic().