[BUGFIX] Exclude unavailable language entries defined in order (#1313)
When "order" is used in v:page.languageMenu and one of the languages from "order" is disabled in backend then its anyway added to $languageMenu array. That leads to problems for example with RealUrl 2.2.0 which disables frontend cache when there is a try to create a link with non existing language.