[TASK] Allow installation of TYPO3 master (#1458)
Add replacement to typo3/cms-version in this package
as this package was removed for TYPO3 8 and merged
into typo3/cms-workspaces.

Simplify travis build by only requiring typo3/cms-core.
This is enough because all TYPO3 packages require
this package in the same version. If typo3/cms-core is required
in dev-master, all other are also installed in dev-master.

This has the additional benefit, that other direct dependencies
to TYPO3 framework packages remain unchanged, thus it is checked
by composer whether they are allowed to be installed in the version
requested for typo3/cms-core.

Add prefer-stable: true to always fetch stable versions when
possible even when composer is called without --prefer-dist.

Last but not least, replace the deprecated or operator "|"
whith the new "||" operator.