[REMOVAL] Remove (broken) support for form object lookup
Affects v:condition.form.hasValidator and v:condition.form.isRequired

This removes the currently broken ability to look for a form
object in the ViewHelperVariableContainer; which has not
worked since refactoring the ViewHelper to static calling.

Also changes the required state of arguments to make sure
a complete set of arguments will always be present.

Constitutes a breaking change if your template does not
pass the “object” argument when calling the two mentioned
ViewHelpers. To work around this, add the “object” argument.

If used in a controller action where the object may be null,
create a condition in the controller action to create a dummy
instance if the instance is NULL, then make sure that gets
assigned to the template and used in the form.