[BUGFIX] Correctly use PageRender meta tag API
Fixes a problem with meta tags created via the
v:page.header.meta ViewHelper on TYPO3 <10.4,
in which a tag added by both the VH and TS/PHP
would not be overwritten. Additionally, fixes an
issue on TYPO3 9.5+ when generating http-equiv
or property meta tags which prevented these from
being possible.