[BUGFIX] Do not process absolute image URIs in the backend
Sometimes the TYPO3v10 backend already returns absolute image URLs that do
not need to be modified anymore.
This happens for a couple of minutes only, afterwards the backend returns relative
URLs.
This patch fixes image URLs in that short timeframe.

Example of such an absolute image URL returned in the backend that needs no
further processing:
 http://localhost:5707/typo3/index.php?route=/image/process&token=414b99de77e22aa555584f4c53fdb5ebb9b2deb9&id=516701