[FEATURE] Use site configuration for language menu
TYPO3 9 includes a new way to configure the languages available on a
site as part of the newly implemented "site" configuration. Among other
things, there is now no special treatment of the default language any
more as far as the configuration is concerned.

vhs, until now, still relied on querying the sys_language table to
gather the languages to display. This is now replaced by querying the
site configuration on supported TYPO3 versions. Although the label
override arguments are still supported on the view helper, they should
now not be necessary any more, since labels to display in the frontend
can be directly specified in the site configuration.