[FEATURE] Use IconFactory for displaying flags in language menu
The flag selection for languages in the site configuration of the TYPO3
backend is not directly based on ISO language codes like the
sys_language table was, but instead uses TYPO3 icon identifiers.
To be more compatible with what the backend uses, vhs can also use the
IconFactory to render the flags. This is, however, only performed if the
user has not otherwise specified a path to flags.