Fixed possible permission problem with cache:flush
The touch command only succeeds if the process user is the file owner. This must not be true in some deployment scenarios: The deployment user is allowed to write (and delete) such cache files but mus not be the owner. The solution is to recreate the cache file (in anatomic operation) with the same contents.