[BUGFIX] Fix PHP 7.3.0 deprecated behavior (#1692)
If needle is not a string, it is converted to an integer and applied as the ordinal value of a character. This behavior is deprecated as of PHP 7.3.0, and relying on it is highly discouraged. Depending on the intended behavior, the needle should either be explicitly cast to string, or an explicit call to chr() should be performed.

Co-authored-by: Claus Due <claus@namelesscoder.net>