[FEATURE] Use site configuration in v:page.languageMenu on TYPO3 9/10  (#1723)
* [FEATURE] Use site configuration for language menu

TYPO3 9 includes a new way to configure the languages available on a
site as part of the newly implemented "site" configuration. Among other
things, there is now no special treatment of the default language any
more as far as the configuration is concerned.

vhs, until now, still relied on querying the sys_language table to
gather the languages to display. This is now replaced by querying the
site configuration on supported TYPO3 versions. Although the label
override arguments are still supported on the view helper, they should
now not be necessary any more, since labels to display in the frontend
can be directly specified in the site configuration.

* [FEATURE] Use IconFactory for displaying flags in language menu

The flag selection for languages in the site configuration of the TYPO3
backend is not directly based on ISO language codes like the
sys_language table was, but instead uses TYPO3 icon identifiers.
To be more compatible with what the backend uses, vhs can also use the
IconFactory to render the flags. This is, however, only performed if the
user has not otherwise specified a path to flags.

Co-authored-by: Claus Due <claus@namelesscoder.net>