[DOCS] Add generated documentation manually the first time
The documentation files were created by manually processing the
"render-docs.yml" on a local machine. This is due to the fact
that the bin/generate-schema binary of the
typo3/fluid-schema-generator project does not yet work properly
with TYPO3 11+, which is documented in
https://github.com/TYPO3-Documentation/t3docs-ci-deploy/issues/4
and has to be fixed to make "render-docs.yml" work properly.