[DOC] Replace "Edit on GitHub" workflow with contribution page
Most of the ViewHelper documentation is generated from the ViewHelper
source code. Therefore, the "Edit on GitHub" workflow is
counterproductive because it applies PRs to the reST files that will
be overwritten during the next generation run. Remove the
"Edit on GitHub" workflow and instead add a contribution page that
explains all the manual steps.

Conflicts:
	Documentation/Index.rst