[BUGFIX] Don't cache TypoScript for pages with uncached content
Since change e84af96c ("[BUGFIX] Handle new TS-not-set limitation on v13
(#1928)") the "plugin.tx_vhs" section of the TypoScript setup is written
into a cache. This is necessary as TYPO3 v13
PrepareTypoScriptFrontendRendering [1] no longer performs the full
TypoScript setup when serving a fully-cached page. An "Setup array has
not been initialized" error is thrown when trying to access the
uninitialized TypoScript setup.

For pages with uncached content, FrontendInterface::set() is called on
each request. This leads to database deadlock errors if the
Typo3DatabaseBackend is used for the "vhs_main" cache and many
concurrent requests hit the same TYPO3 page. This issues is tracked in
upstream TYPO3 bug tracker as #106593 [2]. A fully reproduction example
is available on GitHub [3].

We don't need to store the TypoScript of those pages with uncached
content in the cache, because they cannot be served from cache and thus
force full TypoScript setup even when answering subsequent requests.

[1]: https://github.com/TYPO3/typo3/blob/v13.4.9/typo3/sysext/frontend/Classes/Middleware/PrepareTypoScriptFrontendRendering.php#L146
[2]: https://forge.typo3.org/issues/106593
[3]: https://github.com/adamkoppede/reproduce-deadlock-with-fluidtypo3-vhs-7-1